--- a/ProtoPromise_Unity/Assets/Plugins/ProtoPromiseTests/Tests/Threading/MergeConcurrencyTests.cs
+++ b/ProtoPromise_Unity/Assets/Plugins/ProtoPromiseTests/Tests/Threading/MergeConcurrencyTests.cs
@@ -32,7 +32,7 @@ namespace ProtoPromiseTests.Threading
         [TearDown]
         public void Teardown()
         {
-            Promise.Config.UncaughtRejectionHandler = currentHandler;
+            TestHelper.WaitForAllThreadsAndReplaceRejectionHandler(currentHandler);
 
             TestHelper.Cleanup();
         }