Removed duplicate id checks in awaiters.
Removed a check in AsyncPromiseRef.

Master:

```
|          Method | Pending | Recursion |     Mean |   Error | Allocated | Survived |
|---------------- |-------- |---------- |---------:|--------:|----------:|---------:|
| ProtoPromise_V2 |    True |       100 | 193.2 μs | 1.44 μs |         - | 25,152 B |
```

This PR:

```
|          Method | Pending | Recursion |     Mean |   Error | Allocated | Survived |
|---------------- |-------- |---------- |---------:|--------:|----------:|---------:|
| ProtoPromise_V2 |    True |       100 | 178.6 μs | 1.18 μs |         - | 25,152 B |
```